--- src/instrumentation/LambdaHypertraceInstrumentation.ts
@@ -1,461 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import * as path from 'path';
-
-import {
-    InstrumentationBase,
-    InstrumentationNodeModuleDefinition,
-    InstrumentationNodeModuleFile,
-    isWrapped,
-    safeExecuteInTheMiddle,
-} from '@opentelemetry/instrumentation';
-import {
-    Context as OtelContext,
-    context as otelContext,
-    diag,
-    trace,
-    propagation,
-    Span,
-    SpanKind,
-    SpanStatusCode,
-    TextMapGetter,
-    TraceFlags,
-    TracerProvider,
-    ROOT_CONTEXT,
-} from '@opentelemetry/api';
-import {
-    AWSXRAY_TRACE_ID_HEADER,
-    AWSXRayPropagator,
-} from '@opentelemetry/propagator-aws-xray';
-import {
-    SemanticAttributes,
-    SemanticResourceAttributes,
-} from '@opentelemetry/semantic-conventions';
-
-import {
-    APIGatewayProxyEventHeaders,
-    Callback,
-    Context,
-    Handler,
-} from 'aws-lambda';
-
-import {
-    LambdaModule,
-    AwsLambdaInstrumentationConfig,
-    EventContextExtractor,
-} from '@opentelemetry/instrumentation-aws-lambda/build/src/types';
-import {VERSION} from '@opentelemetry/instrumentation-aws-lambda/build/src/version';
-import * as fs from "fs";
-import {logger} from "../Logging";
-
-const awsPropagator = new AWSXRayPropagator();
-const headerGetter: TextMapGetter<APIGatewayProxyEventHeaders> = {
-    keys(carrier): string[] {
-        return Object.keys(carrier);
-    },
-    get(carrier, key: string) {
-        return carrier[key];
-    },
-};
-
-export const traceContextEnvironmentKey = '_X_AMZN_TRACE_ID';
-const name = '@opentelemetry/instrumentation-aws-lambda'
-
-export class AwsLambdaInstrumentation extends InstrumentationBase {
-    private _forceFlush?: () => Promise<void>;
-
-    constructor(_config: AwsLambdaInstrumentationConfig = {}) {
-        super('@opentelemetry/instrumentation-aws-lambda', VERSION, _config);
-    }
-
-    setConfig(config: AwsLambdaInstrumentationConfig = {}) {
-        this._config = config;
-    }
-
-    init() {
-        const taskRoot = process.env.LAMBDA_TASK_ROOT;
-        const handlerDef = process.env._HANDLER;
-
-        // _HANDLER and LAMBDA_TASK_ROOT are always defined in Lambda but guard bail out if in the future this changes.
-        if (!taskRoot || !handlerDef) {
-            return [];
-        }
-
-        const handler = path.basename(handlerDef);
-        const moduleRoot = handlerDef.substr(0, handlerDef.length - handler.length);
-
-        const [module, functionName] = handler.split('.', 2);
-
-        // Lambda loads user function using an absolute path.
-        let filename = path.resolve(taskRoot, moduleRoot, module);
-        const extensions = ['.js', '.cjs'];
-
-        let fileExists = false;
-        for (let ext of extensions) {
-            try {
-                fs.statSync(`${filename}${ext}`);
-                filename += ext;
-                fileExists = true;
-                break;
-            } catch (_) {
-                // if we couldn't find that file that is okay/expected since we are checking for a few different extensions
-            }
-        }
-
-        if (!fileExists) {
-            return []
-        }
-
-        return [
-            new InstrumentationNodeModuleDefinition(
-                // NB: The patching infrastructure seems to match names backwards, this must be the filename, while
-                // InstrumentationNodeModuleFile must be the module name.
-                filename,
-                ['*'],
-                undefined,
-                undefined,
-                [
-                    new InstrumentationNodeModuleFile(
-                        module,
-                        ['*'],
-                        (moduleExports: LambdaModule) => {
-                            diag.debug('Applying patch for lambda handler');
-                            if (isWrapped(moduleExports[functionName])) {
-                                this._unwrap(moduleExports, functionName);
-                            }
-                            this._wrap(moduleExports, functionName, this._getHandler());
-                            return moduleExports;
-                        },
-                        (moduleExports?: LambdaModule) => {
-                            if (moduleExports == undefined) return;
-                            diag.debug('Removing patch for lambda handler');
-                            this._unwrap(moduleExports, functionName);
-                        }
-                    ),
-                ]
-            ),
-        ];
-    }
-
-    private _getHandler() {
-        return (original: Handler) => {
-            return this._getPatchHandler(original);
-        };
-    }
-
-    private _getPatchHandler(original: Handler) {
-        diag.debug('patch handler function');
-        const plugin = this;
-
-        return function patchedHandler(
-            this: never,
-            // The event can be a user type, it truly is any.
-            // eslint-disable-next-line @typescript-eslint/no-explicit-any
-            event: any,
-            context: Context,
-            callback: Callback
-        ) {
-            const config = plugin._config;
-            const parent = AwsLambdaInstrumentation._determineParent(
-                event,
-                // @ts-ignore
-                config.disableAwsContextPropagation === true,
-                // @ts-ignore
-                config.eventContextExtractor ||
-                AwsLambdaInstrumentation._defaultEventContextExtractor
-            );
-
-            const name = context.functionName;
-            const span = plugin.tracer.startSpan(
-                name,
-                {
-                    kind: SpanKind.SERVER,
-                    attributes: {
-                        [SemanticAttributes.FAAS_EXECUTION]: context.awsRequestId,
-                        [SemanticResourceAttributes.FAAS_ID]: context.invokedFunctionArn,
-                        [SemanticResourceAttributes.CLOUD_ACCOUNT_ID]:
-                            AwsLambdaInstrumentation._extractAccountId(
-                                context.invokedFunctionArn
-                            ),
-                    },
-                },
-                parent
-            );
-
-            // @ts-ignore
-            if (config.requestHook) {
-                safeExecuteInTheMiddle(
-                    // @ts-ignore
-                    () => config.requestHook!(span, {event, context}),
-                    e => {
-                        if (e)
-                            diag.error('aws-lambda instrumentation: requestHook error', e);
-                    },
-                    true
-                );
-            }
-
-            return otelContext.with(trace.setSpan(otelContext.active(), span), () => {
-                // Lambda seems to pass a callback even if handler is of Promise form, so we wrap all the time before calling
-                // the handler and see if the result is a Promise or not. In such a case, the callback is usually ignored. If
-                // the handler happened to both call the callback and complete a returned Promise, whichever happens first will
-                // win and the latter will be ignored.
-                const wrappedCallback = plugin._wrapCallback(callback, span);
-                const maybePromise = safeExecuteInTheMiddle(
-                    () => original.apply(this, [event, context, wrappedCallback]),
-                    error => {
-                        if (error != null) {
-                            // Exception thrown synchronously before resolving callback / promise.
-                            plugin._applyResponseHook(span, error);
-                            plugin._endSpan(span, error, () => {
-                            });
-                        }
-                    }
-                ) as Promise<{}> | undefined;
-                if (typeof maybePromise?.then === 'function') {
-                    return maybePromise.then(
-                        value => {
-                            plugin._applyResponseHook(span, null, value);
-                            return new Promise(resolve =>
-                                plugin._endSpan(span, undefined, () => resolve(value))
-                            );
-                        },
-                        (err: Error | string) => {
-                            plugin._applyResponseHook(span, err);
-                            return new Promise((resolve, reject) =>
-                                plugin._endSpan(span, err, () => reject(err))
-                            );
-                        }
-                    );
-                }
-                return maybePromise;
-            });
-        };
-    }
-
-    setTracerProvider(tracerProvider: TracerProvider) {
-        super.setTracerProvider(tracerProvider);
-        this._forceFlush = AwsLambdaInstrumentation._getForceFlush(tracerProvider);
-    }
-
-    static _getForceFlush(tracerProvider: TracerProvider) {
-        if (!tracerProvider) return undefined;
-
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        let currentProvider: any = tracerProvider;
-
-        if (typeof currentProvider.getDelegate === 'function') {
-            currentProvider = currentProvider.getDelegate();
-        }
-
-        if (typeof currentProvider.forceFlush === 'function') {
-            logger.info("CURRENT PROVIDER", currentProvider)
-            return currentProvider.forceFlush.bind(currentProvider);
-        }
-
-        return undefined;
-    }
-
-    private _wrapCallback(original: Callback, span: Span): Callback {
-        const plugin = this;
-        return function wrappedCallback(this: never, err, res) {
-            diag.debug('executing wrapped lookup callback function');
-            plugin._applyResponseHook(span, err, res);
-
-            plugin._endSpan(span, err, () => {
-                diag.debug('executing original lookup callback function');
-                return original.apply(this, [err, res]);
-            });
-        };
-    }
-
-    private _endSpan(
-        span: Span,
-        err: string | Error | null | undefined,
-        callback: () => void
-    ) {
-        if (err) {
-            span.recordException(err);
-        }
-
-        let errMessage;
-        if (typeof err === 'string') {
-            errMessage = err;
-        } else if (err) {
-            errMessage = err.message;
-        }
-        if (errMessage) {
-            span.setStatus({
-                code: SpanStatusCode.ERROR,
-                message: errMessage,
-            });
-        }
-
-        span.end();
-        if (this._forceFlush) {
-            this._forceFlush().then(
-                () => callback(),
-                () => callback()
-            );
-        } else {
-            diag.error(
-                'Spans may not be exported for the lambda function because we are not force flushing before callback.'
-            );
-            callback();
-        }
-    }
-
-    private _applyResponseHook(
-        span: Span,
-        err?: Error | string | null,
-        res?: any
-    ) {
-        // @ts-ignore
-        if (this._config?.responseHook) {
-            safeExecuteInTheMiddle(
-                // @ts-ignore
-                () => this._config.responseHook!(span, {err, res}),
-                e => {
-                    if (e)
-                        diag.error('aws-lambda instrumentation: responseHook error', e);
-                },
-                true
-            );
-        }
-    }
-
-    private static _extractAccountId(arn: string): string | undefined {
-        const parts = arn.split(':');
-        if (parts.length >= 5) {
-            return parts[4];
-        }
-        return undefined;
-    }
-
-    private static _defaultEventContextExtractor(event: any): OtelContext {
-        // The default extractor tries to get sampled trace header from HTTP headers.
-        const httpHeaders = event.headers || {};
-        return propagation.extract(otelContext.active(), httpHeaders, headerGetter);
-    }
-
-    private static _determineParent(
-        event: any,
-        disableAwsContextPropagation: boolean,
-        eventContextExtractor: EventContextExtractor
-    ): OtelContext {
-        let parent: OtelContext | undefined = undefined;
-        if (!disableAwsContextPropagation) {
-            const lambdaTraceHeader = process.env[traceContextEnvironmentKey];
-            if (lambdaTraceHeader) {
-                parent = awsPropagator.extract(
-                    otelContext.active(),
-                    {[AWSXRAY_TRACE_ID_HEADER]: lambdaTraceHeader},
-                    headerGetter
-                );
-            }
-            if (parent) {
-                const spanContext = trace.getSpan(parent)?.spanContext();
-                if (
-                    spanContext &&
-                    (spanContext.traceFlags & TraceFlags.SAMPLED) === TraceFlags.SAMPLED
-                ) {
-                    // Trace header provided by Lambda only sampled if a sampled context was propagated from
-                    // an upstream cloud service such as S3, or the user is using X-Ray. In these cases, we
-                    // need to use it as the parent.
-                    return parent;
-                }
-            }
-        }
-        const extractedContext = safeExecuteInTheMiddle(
-            () => eventContextExtractor(event),
-            e => {
-                if (e)
-                    diag.error(
-                        'aws-lambda instrumentation: eventContextExtractor error',
-                        e
-                    );
-            },
-            true
-        );
-        if (trace.getSpan(extractedContext)?.spanContext()) {
-            return extractedContext;
-        }
-        if (!parent) {
-            // No context in Lambda environment or HTTP headers.
-            return ROOT_CONTEXT;
-        }
-        return parent;
-    }
-
-    // Until ES Modules are supported with autoinstrumentation
-    // users will manually have to instrument the lambda function
-    public static TraceLambda(handler, config = {}, traceProvider) {
-        const tracer = trace.getTracer(name, VERSION);
-
-        return async function tracedHandler(event, lambdaContext, callback) {
-            const spanName = lambdaContext.functionName || 'Lambda Function';
-            const span = tracer.startSpan(spanName, {
-                kind: SpanKind.SERVER,
-                attributes: {
-                    [SemanticAttributes.FAAS_EXECUTION]: lambdaContext.awsRequestId,
-                    [SemanticResourceAttributes.FAAS_ID]: lambdaContext.invokedFunctionArn,
-                },
-            });
-
-            // @ts-ignore
-            if (config.requestHook) {
-                // @ts-ignore
-                config.requestHook(span, {event, context: lambdaContext});
-            }
-
-            // Use the OpenTelemetry context API to set the current span as active
-            return otelContext.with(trace.setSpan(otelContext.active(), span), async () => {
-                try {
-                    // Execute the original handler function
-                    const result = await handler(event, lambdaContext, callback);
-
-                    // @ts-ignore
-                    if (config.responseHook) {
-                        // @ts-ignore
-                        config.responseHook(span, {err: null, res: result});
-                    }
-
-                    span.end();
-                    await traceProvider.forceFlush()
-
-                    return result;
-                } catch (error) {
-                    // @ts-ignore
-                    if (config.responseHook) {
-                        // @ts-ignore
-                        config.responseHook(span, {err: error});
-                    }
-
-                    span.recordException(error);
-                    span.setStatus({
-                        code: SpanStatusCode.ERROR,
-                        message: error.message,
-                    });
-                    span.end();
-                    await traceProvider.forceFlush()
-                    throw error; // Rethrow the error to ensure Lambda can handle it accordingly
-                }
-            });
-        };
-    }
-}
-
